scripts/requirements.txt
numpy>=1.24
torch>=2.0
faiss-gpu>=1.7.4
sentence-transformers>=2.2
